+++ src/Charta/Metadata/XmpWriter.cs
@@ -10,13 +10,13 @@ namespace Charta.Metadata;
 /// </summary>
 internal static class XmpWriter
 {
-    public static byte[] Build(DocumentMetadata metadata) => Build(metadata, pdfAConformance: null);
+    public static byte[] Build(DocumentMetadata metadata) => Build(metadata, pdfAConformance: null, pdfUa: false);
 
     /// <summary>
-    /// Builds the XMP packet. When <paramref name="pdfAConformance"/> is set (e.g. "2B"), the pdfaid
-    /// schema is included so validators recognize the PDF/A level.
+    /// Builds the XMP packet. When <paramref name="pdfAConformance"/> is set (e.g. "2B") the pdfaid
+    /// schema is included; when <paramref name="pdfUa"/> is set the pdfuaid schema (part 1) is too.
     /// </summary>
-    public static byte[] Build(DocumentMetadata metadata, string? pdfAConformance)
+    public static byte[] Build(DocumentMetadata metadata, string? pdfAConformance, bool pdfUa)
     {
         var sb = new StringBuilder();
         sb.Append("<?xpacket begin=\"\uFEFF\" id=\"W5M0MpCehiHzreSzNTczkc9d\"?>\n");
@@ -26,7 +26,10 @@ internal static class XmpWriter
         sb.Append("    xmlns:dc=\"http://purl.org/dc/elements/1.1/\"\n");
         sb.Append("    xmlns:xmp=\"http://ns.adobe.com/xap/1.0/\"\n");
         sb.Append("    xmlns:pdf=\"http://ns.adobe.com/pdf/1.3/\"\n");
-        sb.Append("    xmlns:pdfaid=\"http://www.aiim.org/pdfa/ns/id/\">\n");
+        sb.Append("    xmlns:pdfaid=\"http://www.aiim.org/pdfa/ns/id/\"");
+        sb.Append(pdfUa
+            ? "\n    xmlns:pdfuaid=\"http://www.aiim.org/pdfua/ns/id/\">\n"
+            : ">\n");
 
         if (pdfAConformance is not null)
         {
@@ -34,6 +37,11 @@ internal static class XmpWriter
             var conformance = pdfAConformance[^1..].ToUpperInvariant();
             sb.Append("   <pdfaid:part>").Append(part).Append("</pdfaid:part>\n");
             sb.Append("   <pdfaid:conformance>").Append(conformance).Append("</pdfaid:conformance>\n");
+        }
+
+        if (pdfUa)
+        {
+            sb.Append("   <pdfuaid:part>1</pdfuaid:part>\n");
         }
 
         if (metadata.Title is { } title)